refactor: split god-hook and god-component, fix version switching layer violations
- Split useGenerateForm (1232 → 210 lines) into 4 focused hooks:
  useGenerateFormTypes, useGenerateFormDraft, useGenerateFormValidation,
  useGenerateFormSubmit
- Split GlobalAudioProvider (982 → 395 lines) into 3 hooks:
  useAudioSourceResolver, usePlaybackStatusSync, useAudioErrorRecovery
- Fix version switching in 3 components to use useVersionSwitcher hook
  instead of raw Supabase calls (layer violation fix)
- Update PROJECT_STATUS.md with progress
