+++ src/components/GlobalAudioProvider.tsx
@@ -13,12 +13,10 @@
  * - Caching: Integrates with IndexedDB audio cache for offline playback
  * - Prefetching: Automatically prefetches next tracks in queue
  *
- * @optimizations
- * 1. Debounced time updates: Reduces re-renders from 60fps to ~10fps
- * 2. Audio prefetching: Prefetches next 2 tracks in queue for instant playback
- * 3. Enhanced error handling: User-friendly error messages with recovery options
- * 4. Position persistence: Saves/restores playback position across sessions
- * 5. Startup period suppression: Prevents stale data toasts during initial load
+ * Hooks extracted for modularity:
+ * - usePlaybackStatusSync: audio event → store status sync
+ * - useAudioSourceResolver: URL resolution with mobile format filtering
+ * - useAudioErrorRecovery: error handling, stall recovery, auto-skip
  *
  * @usage
  * ```tsx
@@ -49,104 +47,33 @@ import { setGlobalAudioRef } from "@/hooks/audio/useAudioTime";
 import { useOptimizedAudioPlayer } from "@/hooks/audio/useOptimizedAudioPlayer";
 import { usePlaybackPosition } from "@/hooks/audio/usePlaybackPosition";
 import { usePlayerAnalytics } from "@/hooks/audio/usePlayerAnalytics";
+import { usePlaybackStatusSync } from "@/hooks/audio/usePlaybackStatusSync";
+import { useAudioSourceResolver } from "@/hooks/audio/useAudioSourceResolver";
+import { useAudioErrorRecovery } from "@/hooks/audio/useAudioErrorRecovery";
 import { logger } from "@/lib/logger";
 import { toast } from "@/lib/toast";
-import { playerAnalytics, recordError } from "@/lib/telemetry";
-import { detectMobileBrowser, isAudioFormatSupported, logAudioDiagnostics } from "@/lib/audioFormatUtils";
-
-// Audio error messages by error code with detailed recovery info
-const AUDIO_ERROR_MESSAGES: Record<
-  number,
-  {
-    ru: string;
-    action?: string;
-    retryable?: boolean;
-    errorType: string;
-  }
-> = {
-  1: {
-    ru: "Загрузка аудио прервана",
-    action: "Попробуйте еще раз",
-    retryable: true,
-    errorType: "MEDIA_ERR_ABORTED",
-  },
-  2: {
-    ru: "Сетевая ошибка при загрузке",
-    action: "Проверьте подключение",
-    retryable: true,
-    errorType: "MEDIA_ERR_NETWORK", // Includes PIPELINE_ERROR_READ
-  },
-  3: {
-    ru: "Ошибка декодирования аудио",
-    action: "Файл может быть поврежден",
-    retryable: false,
-    errorType: "MEDIA_ERR_DECODE",
-  },
-  4: {
-    ru: "Формат аудио не поддерживается",
-    action: "Попробуйте другой трек",
-    retryable: true,
-    errorType: "MEDIA_ERR_SRC_NOT_SUPPORTED",
-  },
-};
+import { playerAnalytics } from "@/lib/telemetry";
+import { logAudioDiagnostics } from "@/lib/audioFormatUtils";
 
 export function GlobalAudioProvider({ children }: { children: React.ReactNode }) {
   const audioRef = useRef<HTMLAudioElement | null>(null);
   const lastTrackIdRef = useRef<string | null>(null);
   // Suppress errors during initial startup to avoid stale data toasts
   const mountTimeRef = useRef(Date.now());
-  const isStartupPeriod = () => Date.now() - mountTimeRef.current < 2000;
+  const isStartupPeriod = useCallback(() => Date.now() - mountTimeRef.current < 2000, []);
 
-  // Detect mobile browser for enhanced error handling
-  const mobileBrowserInfo = useRef(detectMobileBrowser());
-  const isMobileBrowser = mobileBrowserInfo.current.isMobile;
+  const { activeTrack, isPlaying, volume, pauseTrack } = usePlayerStore();
 
-  const { activeTrack, isPlaying, repeat, volume, pauseTrack, nextTrack } = usePlayerStore();
-  const setPlaybackStatus = usePlayerStore((s) => s.setPlaybackStatus);
+  // --- Extracted hooks ---
 
-  // Sync granular playback status from <audio> events → store, so UI can
-  // render distinct loading / buffering / error states (not just play/pause).
-  useEffect(() => {
-    const audio = audioRef.current;
-    if (!audio) return;
+  // Sync granular playback status from <audio> events → store
+  usePlaybackStatusSync(audioRef);
 
-    const onLoadStart = () => setPlaybackStatus(activeTrack ? "loading" : "idle");
-    const onWaiting = () => setPlaybackStatus("buffering");
-    const onPlaying = () => setPlaybackStatus("playing");
-    const onPause = () => {
-      // Don't override 'loading'/'buffering' with 'paused' if we never started.
-      if (audio.currentTime > 0 || audio.readyState >= 2) setPlaybackStatus("paused");
-    };
-    const onEnded = () => setPlaybackStatus("idle");
-    const onCanPlay = () => {
-      if (audio.paused) setPlaybackStatus("paused");
-    };
-    const onErr = () => {
-      const msg = audio.error?.message || AUDIO_ERROR_MESSAGES[audio.error?.code || 0]?.ru || "Ошибка воспроизведения";
-      setPlaybackStatus("error", msg);
-    };
-
-    audio.addEventListener("loadstart", onLoadStart);
-    audio.addEventListener("waiting", onWaiting);
-    audio.addEventListener("playing", onPlaying);
-    audio.addEventListener("pause", onPause);
-    audio.addEventListener("ended", onEnded);
-    audio.addEventListener("canplay", onCanPlay);
-    audio.addEventListener("error", onErr);
-
-    return () => {
-      audio.removeEventListener("loadstart", onLoadStart);
-      audio.removeEventListener("waiting", onWaiting);
-      audio.removeEventListener("playing", onPlaying);
-      audio.removeEventListener("pause", onPause);
-      audio.removeEventListener("ended", onEnded);
-      audio.removeEventListener("canplay", onCanPlay);
-      audio.removeEventListener("error", onErr);
-    };
-  }, [activeTrack, setPlaybackStatus]);
+  // Resolve audio source URL with mobile format filtering
+  const { getAudioSource, isMobileBrowser, mobileBrowserInfo } = useAudioSourceResolver({ isStartupPeriod });
 
   // Use optimized audio player with caching and prefetch
-  const { prefetchNextTracks } = useOptimizedAudioPlayer({
+  useOptimizedAudioPlayer({
     enablePrefetch: true,
     enableCache: true,
     crossfadeDuration: 0.3,
@@ -157,6 +84,23 @@ export function GlobalAudioProvider({ children }: { children: React.ReactNode })
 
   // Use player analytics tracking
   usePlayerAnalytics();
+
+  // Stable ref for isPlaying to avoid effect re-runs
+  const isPlayingRef = useRef(isPlaying);
+  isPlayingRef.current = isPlaying;
+
+  // Ref to track if we're currently loading a new track
+  const isLoadingRef = useRef(false);
+  const playPromiseRef = useRef<Promise<void> | null>(null);
+
+  // Error recovery (ended, error, stalled, suspend)
+  useAudioErrorRecovery({
+    audioRef,
+    playPromiseRef,
+    isMobileBrowser,
+    mobileBrowserInfo,
+    isStartupPeriod,
+  });
 
   // Initialize audio element once and clean up invalid tracks from localStorage
   useEffect(() => {
@@ -234,147 +178,6 @@ export function GlobalAudioProvider({ children }: { children: React.ReactNode })
     }
   }, [volume]);
 
-  // Get audio source from track with validation and detailed logging
-  const getAudioSource = useCallback(() => {
-    if (!activeTrack) {
-      logger.debug("No active track");
-      return null;
-    }
-
-    // Build fallback chain: prefer streaming_url > local_audio_url > audio_url
-    // On mobile with format errors, we'll skip blob URLs in favor of canonical URLs
-    const sources = [
-      { url: activeTrack.local_audio_url, label: "local_audio_url (blob)" },
-      { url: activeTrack.streaming_url, label: "streaming_url" },
-      { url: activeTrack.audio_url, label: "audio_url" },
-    ].filter((s) => s.url); // Remove null/undefined sources
-
-    // If no sources available, show error
-    if (sources.length === 0) {
-      logger.warn("Track has no audio source", {
-        trackId: activeTrack.id,
-        title: activeTrack.title,
-        status: activeTrack.status,
-      });
-
-      // Show user-friendly error
-      toast.error("Трек не готов к воспроизведению", {
-        description:
-          activeTrack.status === "processing" ? "Трек еще генерируется, подождите..." : "Файл трека отсутствует",
-      });
-
-      return null;
-    }
-
-    // On mobile devices, proactively check format compatibility
-    if (isMobileBrowser) {
-      // Filter out sources that are known to be incompatible
-      const compatibleSources = sources.filter((s) => {
-        const url = s.url!;
-
-        // Skip blob URLs on mobile if we have canonical URLs available
-        // This prevents format errors that would need recovery
-        if (url.startsWith("blob:") && sources.length > 1) {
-          logger.debug("Skipping blob URL on mobile in favor of canonical URL", {
-            trackId: activeTrack.id,
-            availableSources: sources.length,
-          });
-          return false;
-        }
-
-        // Check format compatibility for non-blob URLs
-        if (!url.startsWith("blob:")) {
-          const isSupported = isAudioFormatSupported(url);
-          if (!isSupported) {
-            logger.warn("Audio format may not be supported on mobile", {
-              trackId: activeTrack.id,
-              url: url.substring(0, 60),
-              browser: mobileBrowserInfo.current.browserName,
-            });
-          }
-          return isSupported;
-        }
-
-        return true;
-      });
-
-      // If we filtered everything out, use original list (let browser try)
-      const finalSources = compatibleSources.length > 0 ? compatibleSources : sources;
-
-      // Use first compatible source
-      const selectedSource = finalSources[0];
-      logger.info("Selected audio source for mobile", {
-        trackId: activeTrack.id,
-        selectedSource: selectedSource.label,
-        url: selectedSource.url!.substring(0, 60),
-        totalSources: sources.length,
-        compatibleSources: compatibleSources.length,
-      });
-
-      return selectedSource.url!;
-    }
-
-    // On desktop, use first available source
-    const source = sources[0].url!;
-
-    // Check for valid URL format
-    try {
-      // For blob URLs, just check prefix
-      if (source.startsWith("blob:")) {
-        logger.debug("Using blob URL", { trackId: activeTrack.id });
-        return source;
-      }
-
-      // For data URLs, check format
-      if (source.startsWith("data:")) {
-        if (source.startsWith("data:audio/")) {
-          logger.debug("Using data URL", { trackId: activeTrack.id });
-          return source;
-        }
-        logger.warn("Invalid data URL format", { trackId: activeTrack.id });
-        return null;
-      }
-
-      // For HTTP(S) URLs, validate
-      const url = new URL(source);
-      if (url.protocol !== "http:" && url.protocol !== "https:") {
-        logger.warn("Invalid audio URL protocol", { protocol: url.protocol, trackId: activeTrack.id });
-        // Don't show toast during startup to avoid stale data errors
-        if (!isStartupPeriod()) {
-          toast.error("Неверный формат URL аудио");
-        }
-        return null;
-      }
-
-      logger.debug("Using HTTP(S) URL", {
-        trackId: activeTrack.id,
-        protocol: url.protocol,
-        hostname: url.hostname,
-      });
-      return source;
-    } catch (err) {
-      logger.error("Invalid audio URL", err instanceof Error ? err : new Error(String(err)), {
-        source: source.substring(0, 100),
-        trackId: activeTrack.id,
-      });
-      // Don't show toast during startup to avoid stale data errors
-      if (!isStartupPeriod()) {
-        toast.error("Ошибка URL аудио", {
-          description: "Неверный формат ссылки на аудиофайл",
-        });
-      }
-      return null;
-    }
-  }, [activeTrack, isMobileBrowser, mobileBrowserInfo, isStartupPeriod]);
-
-  // Stable ref for isPlaying to avoid effect re-runs
-  const isPlayingRef = useRef(isPlaying);
-  isPlayingRef.current = isPlaying;
-
-  // Ref to track if we're currently loading a new track
-  const isLoadingRef = useRef(false);
-  const playPromiseRef = useRef<Promise<void> | null>(null);
-
   // Effect for loading new tracks - only triggers on track change
   useEffect(() => {
     const audio = audioRef.current;
@@ -448,7 +251,6 @@ export function GlobalAudioProvider({ children }: { children: React.ReactNode })
       // CRITICAL: Wait for loading to complete before playing
       if (isLoadingRef.current) {
         logger.debug("Waiting for audio to load before playing");
-        // Wait for loadeddata or canplaythrough
         const waitForLoad = new Promise<void>((resolve) => {
           const checkReady = () => {
             if (!isLoadingRef.current || audio.readyState >= 2) {
@@ -520,7 +322,6 @@ export function GlobalAudioProvider({ children }: { children: React.ReactNode })
       } catch (error: unknown) {
         const err = error as { name?: string };
         if (err.name === "AbortError" || isCleanedUp) {
-          // Ignore abort errors from track changes
           return;
         }
 
@@ -530,14 +331,12 @@ export function GlobalAudioProvider({ children }: { children: React.ReactNode })
         });
 
         if (err.name === "NotAllowedError") {
-          // Don't show toast during startup to avoid stale data errors
           if (!isStartupPeriod()) {
             toast.error("Воспроизведение заблокировано", {
               description: "Нажмите на экран и попробуйте снова",
             });
           }
         } else if (err.name === "NotSupportedError") {
-          // Don't show toast during startup to avoid stale data errors
           if (!isStartupPeriod()) {
             toast.error("Формат аудио не поддерживается");
           }
@@ -590,393 +389,7 @@ export function GlobalAudioProvider({ children }: { children: React.ReactNode })
       isCleanedUp = true;
       if (playTimeoutId) clearTimeout(playTimeoutId);
     };
-  }, [isPlaying, activeTrack?.id, volume, pauseTrack]);
-
-  // Handle track ended and errors with retry logic
-  useEffect(() => {
-    const audio = audioRef.current;
-    if (!audio) return;
-
-    const handleEnded = () => {
-      logger.debug("Track ended", { trackId: activeTrack?.id });
-      playerAnalytics.trackComplete(activeTrack?.id || "", audio.currentTime, audio.duration);
-      if (repeat === "one") {
-        // Ensure audio is still available and valid before repeating
-        if (audio.src && audio.duration > 0) {
-          audio.currentTime = 0;
-          // Only play if we're still in playing state
-          if (isPlaying) {
-            audio.play().catch((err) => {
-              logger.warn("Repeat play failed", err);
-              // If repeat play fails, try moving to next track
-              if (err.name !== "AbortError") {
-                nextTrack();
-              }
-            });
-          }
-        } else {
-          // Source is invalid, move to next track
-          logger.warn("Cannot repeat track: invalid source");
-          nextTrack();
-        }
-      } else {
-        nextTrack();
-      }
-    };
-
-    // Track retry attempts for failed loads - separate counters for different error types
-    let networkRetryCount = 0;
-    let formatRetryCount = 0;
-    const MAX_NETWORK_RETRIES = 3;
-    const MAX_FORMAT_RETRIES = 2;
-    let retryTimeoutId: NodeJS.Timeout | null = null;
-
-    // Track if we've attempted blob recovery to prevent loops
-    let hasAttemptedBlobRecovery = false;
-
-    // Track attempted URLs to avoid retry loops
-    const attemptedUrls = new Set<string>();
-
-    const handleError = () => {
-      // Ignore errors when src is empty or not set (prevents "Empty src attribute" spam)
-      const currentSrc = audio.src || "";
-      if (!currentSrc || currentSrc === "" || currentSrc === window.location.href) {
-        logger.debug("Ignoring error for empty/invalid src");
-        return;
-      }
-
-      // During startup period, only log but don't record to telemetry
-      // This prevents thousands of errors from stale localStorage data
-      const inStartupPeriod = isStartupPeriod();
-
-      const errorCode = audio.error?.code || 0;
-      const errorInfo = AUDIO_ERROR_MESSAGES[errorCode] || {
-        ru: "Ошибка воспроизведения",
-      };
-
-      const isBlobSource = audio.src?.startsWith("blob:");
-      const canonicalUrl = activeTrack?.streaming_url || activeTrack?.audio_url;
-
-      // Determine if this is a PIPELINE_ERROR_READ (common FFmpeg/Chromium error)
-      const isPipelineError =
-        audio.error?.message?.includes("PIPELINE_ERROR_READ") || audio.error?.message?.includes("FFmpegDemuxer");
-
-      // Enhanced error context with comprehensive debugging info
-      const errorContext = {
-        errorCode,
-        errorType: errorInfo.errorType,
-        errorMessage: audio.error?.message,
-        isPipelineError,
-        trackId: activeTrack?.id,
-        title: activeTrack?.title,
-        source: audio.src?.substring(0, 100),
-        isBlobSource,
-        hasStreamingUrl: !!activeTrack?.streaming_url,
-        hasAudioUrl: !!activeTrack?.audio_url,
-        hasLocalAudioUrl: !!activeTrack?.local_audio_url,
-        networkRetryCount,
-        formatRetryCount,
-        isMobile: isMobileBrowser,
-        browser: mobileBrowserInfo.current.browserName,
-        os: mobileBrowserInfo.current.osName,
-        readyState: audio.readyState,
-        networkState: audio.networkState,
-        bufferedRanges: audio.buffered.length,
-        duration: audio.duration,
-        currentTime: audio.currentTime,
-      };
-
-      // During startup, suppress toasts to avoid stale data errors
-      const suppressToast = inStartupPeriod;
-
-      // Phase 2.1: Enhanced network error handling (code 2) with cache-busting retry
-      // PIPELINE_ERROR_READ and similar network errors should try alternative URLs
-      if (errorCode === 2 && networkRetryCount < MAX_NETWORK_RETRIES) {
-        networkRetryCount++;
-
-        // Build fallback chain for network errors
-        const allAudioUrls = [activeTrack?.streaming_url, activeTrack?.audio_url, activeTrack?.local_audio_url].filter(
-          (url) => url && !attemptedUrls.has(url),
-        );
-
-        // If we have untried URLs, try them
-        if (allAudioUrls.length > 0) {
-          const fallbackUrl = allAudioUrls[0]!;
-          attemptedUrls.add(fallbackUrl);
-
-          logger.info("Network error (code 2), trying alternative URL", {
-            ...errorContext,
-            fallbackUrl: fallbackUrl.substring(0, 60),
-            attempt: networkRetryCount,
-            maxRetries: MAX_NETWORK_RETRIES,
-          });
-
-          const currentTime = audio.currentTime;
-          const wasPlaying = isPlaying;
-
-          audio.src = fallbackUrl;
-          audio.load();
-
-          audio.addEventListener(
-            "canplay",
-            async function onCanPlay() {
-              audio.removeEventListener("canplay", onCanPlay);
-
-              if (currentTime > 0 && !isNaN(currentTime)) {
-                audio.currentTime = currentTime;
-              }
-
-              if (wasPlaying) {
-                try {
-                  await audio.play();
-                  logger.info("✅ Playback recovered after network error with alternative URL", {
-                    trackId: activeTrack?.id,
-                    fallbackUrl: fallbackUrl.substring(0, 60),
-                  });
-                } catch (playErr) {
-                  logger.warn("Recovery play after network error failed", {
-                    error: playErr instanceof Error ? playErr.message : String(playErr),
-                  });
-                }
-              }
-            },
-            { once: true },
-          );
-
-          return;
-        }
-
-        // No alternative URLs, try cache-busting retry
-        const currentUrl = audio.src;
-        if (currentUrl && !currentUrl.includes("retry=")) {
-          const urlSeparator = currentUrl.includes("?") ? "&" : "?";
-          const cacheBustingUrl = `${currentUrl}${urlSeparator}retry=${Date.now()}`;
-
-          logger.info("Network error (code 2), retrying with cache-busting", {
-            ...errorContext,
-            attempt: networkRetryCount,
-            maxRetries: MAX_NETWORK_RETRIES,
-          });
-
-          // Exponential backoff: 1s, 2s, 4s
-          const retryDelay = Math.pow(2, networkRetryCount - 1) * 1000;
-
-          retryTimeoutId = setTimeout(() => {
-            audio.src = cacheBustingUrl;
-            audio.load();
-
-            if (isPlaying) {
-              audio.play().catch((playErr) => {
-                logger.warn("Cache-busting retry play failed", playErr);
-              });
-            }
-
-            if (!suppressToast) {
-              toast.info("Повторная попытка загрузки...", {
-                description: `Попытка ${networkRetryCount} из ${MAX_NETWORK_RETRIES}`,
-              });
-            }
-          }, retryDelay);
-
-          return;
-        }
-      }
-
-      // CRITICAL: Handle format error (code 4) with automatic recovery using fallback chain
-      // DO NOT log to Sentry until all recovery attempts fail
-      if (errorCode === 4 && !hasAttemptedBlobRecovery && formatRetryCount < MAX_FORMAT_RETRIES) {
-        hasAttemptedBlobRecovery = true;
-        formatRetryCount++;
-
-        // Build fallback chain with all available audio URLs
-        // Step 1: Collect all possible URLs (including local_audio_url)
-        const allAudioUrls = [activeTrack?.streaming_url, activeTrack?.audio_url, activeTrack?.local_audio_url].filter(
-          (url) => url && !attemptedUrls.has(url),
-        ); // Remove null/undefined and already tried URLs
-
-        // Step 2: Deduplicate URLs to avoid redundant retry attempts
-        const uniqueUrls = Array.from(new Set(allAudioUrls));
-
-        // Step 3: Filter out the current failed URL
-        let fallbackChain = uniqueUrls.filter((url) => url !== audio.src);
-
-        // Step 4: CRITICAL FIX - If fallback chain is empty (all URLs are the same),
-        // add a cache-busting retry of the original URL
-        // This handles cases where streaming_url, audio_url, and local_audio_url all point to the same resource
-        let isRetryingSameUrl = false;
-        if (fallbackChain.length === 0 && uniqueUrls.length > 0) {
-          // Extract the base URL without existing query params
-          const baseUrl = uniqueUrls[0]!;
-          const urlSeparator = baseUrl.includes("?") ? "&" : "?";
-          const cacheBustingUrl = `${baseUrl}${urlSeparator}retry=${Date.now()}`;
-          fallbackChain = [cacheBustingUrl];
-          isRetryingSameUrl = true;
-
-          logger.info("All audio URLs are identical, retrying with cache-busting", {
-            trackId: activeTrack?.id,
-            originalUrl: baseUrl.substring(0, 60),
-            uniqueUrlsCount: uniqueUrls.length,
-            isMobile: isMobileBrowser,
-          });
-        }
-
-        if (fallbackChain.length > 0) {
-          const fallbackUrl = fallbackChain[0]!;
-          attemptedUrls.add(fallbackUrl);
-
-          // Log recovery attempt (info level, not error)
-          logger.info("Format error (code 4), attempting fallback to next URL in chain", {
-            ...errorContext,
-            currentSource: isBlobSource ? "blob URL" : "canonical URL",
-            fallbackUrl: fallbackUrl?.substring(0, 60),
-            fallbacksRemaining: fallbackChain.length,
-            recoveryAttempt: true,
-            isRetryingSameUrl,
-            strategy: isRetryingSameUrl ? "cache-busting-retry" : "alternative-url",
-          });
-
-          // Save current time before switching source
-          const currentTime = audio.currentTime;
-          const wasPlaying = isPlaying;
-
-          audio.src = fallbackUrl;
-          audio.load();
-
-          // Restore position and resume playback when ready
-          audio.addEventListener(
-            "canplay",
-            async function onCanPlay() {
-              audio.removeEventListener("canplay", onCanPlay);
-
-              if (currentTime > 0 && !isNaN(currentTime)) {
-                audio.currentTime = currentTime;
-              }
-
-              if (wasPlaying) {
-                try {
-                  await audio.play();
-                  logger.info("✅ Playback recovered successfully after format error", {
-                    trackId: activeTrack?.id,
-                    isMobile: isMobileBrowser,
-                    fallbackUsed: fallbackUrl?.substring(0, 60),
-                    wasRetryingSameUrl: isRetryingSameUrl,
-                  });
-                } catch (playErr) {
-                  // Recovery failed - NOW log to Sentry
-                  logger.error("❌ Recovery play after format error failed", playErr, {
-                    ...errorContext,
-                    fallbackUrl: fallbackUrl?.substring(0, 60),
-                    wasRetryingSameUrl: isRetryingSameUrl,
-                  });
-                  recordError(
-                    `audio:${errorCode}:recovery_failed`,
-                    audio.error?.message || "Recovery failed after format error",
-                    {
-                      ...errorContext,
-                      fallbackUrl: fallbackUrl?.substring(0, 60),
-                      wasRetryingSameUrl: isRetryingSameUrl,
-                    },
-                  );
-                }
-              }
-            },
-            { once: true },
-          );
-
-          return; // Don't show error toast or report to Sentry for recoverable format errors
-        } else {
-          // No fallbacks available - log to Sentry
-          logger.error("❌ Format error (code 4) with no fallback URLs available", null, errorContext);
-        }
-      }
-
-      // Log error with full context
-      logger.error("Audio playback error", null, errorContext);
-
-      // Record error in telemetry ONLY if not in startup period
-      // This prevents thousands of errors from stale localStorage data
-      if (!inStartupPeriod) {
-        // Use only one telemetry call to avoid duplicate entries
-        recordError(`audio:${errorCode}`, audio.error?.message || "Unknown audio error", errorContext);
-      }
-
-      // Max retries reached or non-retryable error
-      // Show user-friendly error message (but not during startup)
-      if (!suppressToast) {
-        toast.error(errorInfo.ru, {
-          description: errorInfo.action,
-        });
-      }
-
-      pauseTrack();
-
-      // Auto-skip to next track after 2 seconds for better UX
-      retryTimeoutId = setTimeout(() => {
-        logger.debug("Auto-skipping to next track after error");
-        nextTrack();
-      }, 2000);
-    };
-
-    const handleStalled = async () => {
-      logger.warn("Audio playback stalled", {
-        trackId: activeTrack?.id,
-        currentTime: audio.currentTime,
-        buffered: audio.buffered.length,
-      });
-
-      // CRITICAL FIX: Wait for any pending play promise before attempting recovery
-      // This prevents AbortError: "The play() request was interrupted by a new load request"
-      if (playPromiseRef.current) {
-        logger.debug("Waiting for pending play promise before stall recovery");
-        try {
-          await playPromiseRef.current;
-        } catch (err) {
-          // Ignore errors from the pending play promise
-          logger.debug(
-            "Pending play promise rejected during stall recovery",
-            err instanceof Error ? { message: err.message } : undefined,
-          );
-        }
-        playPromiseRef.current = null;
-      }
-
-      // Try to recover by reloading
-      const currentTime = audio.currentTime;
-      audio.load();
-      audio.currentTime = currentTime;
-
-      if (isPlaying) {
-        audio.play().catch((err) => {
-          logger.error("Failed to recover from stall", err);
-        });
-      }
-    };
-
-    const handleSuspend = () => {
-      logger.debug("Audio loading suspended", {
-        trackId: activeTrack?.id,
-        networkState: audio.networkState,
-        readyState: audio.readyState,
-      });
-    };
-
-    audio.addEventListener("ended", handleEnded);
-    audio.addEventListener("error", handleError);
-    audio.addEventListener("stalled", handleStalled);
-    audio.addEventListener("suspend", handleSuspend);
-
-    return () => {
-      audio.removeEventListener("ended", handleEnded);
-      audio.removeEventListener("error", handleError);
-      audio.removeEventListener("stalled", handleStalled);
-      audio.removeEventListener("suspend", handleSuspend);
-
-      // Clear retry timeout on cleanup
-      if (retryTimeoutId) {
-        clearTimeout(retryTimeoutId);
-      }
-    };
-  }, [repeat, nextTrack, pauseTrack, activeTrack, isPlaying]);
+  }, [isPlaying, activeTrack?.id, volume, pauseTrack, isStartupPeriod]);
 
   return <>{children}</>;
 }
